feat: merge release 2.2.0
* feat: restructurize project

* feat: extract messages and callbacks to the separate projects

* feat: rename web folder

* feat: implement second version

* feat: add first private chat message handlers and infrastructure for bot's second version

* feat: add callbacks

* feat: add callback handler for enqueueing

* refactor: refactor message handlers

* refactor: refactor some callback handlers

* refactor: refactor second part of callback handlers (excluding one)

* feat: add dequeue and remove callback handlers

* feat: make GetAvailablePositions work for exception situations and add unit tests

* feat: add unit tests

* feat: add return buttons to chat [not working version]

* refactor: change project structure and refactor callbacks passing commands as json

* refactor: minimize json size and fix errors

* feat: add exceptions handling middleware; remove message parsing in callback and message handling exceptions handlers

* refactor: refactor project repository

* refactor: move configuration to data and make minor refactoring

* refactor: move all extensions from utilities to messages. Remove utilities project

* feat: add ability to remove queue by chat admin from private chats

+++ Enqueuer.Callbacks/CallbackHandlers/RemoveQueueCallbackHandler.cs
@@ -1,9 +1,11 @@
 ﻿using System.Threading.Tasks;
 using Enqueuer.Callbacks.CallbackHandlers.BaseClasses;
 using Enqueuer.Callbacks.Exceptions;
+using Enqueuer.Callbacks.Extensions;
 using Enqueuer.Data;
 using Enqueuer.Data.Constants;
 using Enqueuer.Data.DataSerialization;
+using Enqueuer.Persistence.Extensions;
 using Enqueuer.Persistence.Models;
 using Enqueuer.Services.Interfaces;
 using Telegram.Bot;
@@ -78,7 +80,18 @@ namespace Enqueuer.Callbacks.CallbackHandlers
         {
             if (callbackData.QueueData.IsUserAgreed.Value)
             {
+                var userId = callbackQuery.From.Id;
                 var chat = queue.Chat;
+                if (!queue.IsQueueCreator(userId) && !await botClient.IsChatAdmin(userId, chat.ChatId))
+                {
+                    return await botClient.EditMessageTextAsync(
+                        callbackQuery.Message.Chat,
+                        callbackQuery.Message.MessageId,
+                        $"Unable to delete <b>'{queue.Name}'</b> queue: you are not queue creator or the chat admin.",
+                        ParseMode.Html,
+                        replyMarkup: this.GetReturnToQueueButton(callbackData));
+                }
+
                 await this.queueService.DeleteQueueAsync(queue);
 
                 await botClient.SendTextMessageAsync(
